+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/OnlineIndexer.java
@@ -140,7 +140,7 @@ public class OnlineIndexer implements AutoCloseable {
     public static final int UNLIMITED = Integer.MAX_VALUE;
 
     /**
-     * If {@link Builder#getIncreaseLimitAfter()} is this value, the limit will not go back up, no matter how many
+     * If {@link OnlineIndexer.Builder#getIncreaseLimitAfter()} is this value, the limit will not go back up, no matter how many
      * successes there are.
      * This is the default value.
      */
@@ -162,16 +162,15 @@ public class OnlineIndexer implements AutoCloseable {
     @Nonnull private final Collection<RecordType> recordTypes;
     @Nonnull private final TupleRange recordsRange;
 
-    private final int maxLimit;
-    private final int maxRetries;
-    private final int recordsPerSecond;
-    private final long progressLogIntervalMillis;
-    private final int increaseLimitAfter;
     /**
      * The current number of records to process in a single transaction, this may go up or down when using
-     * {@link #runAsync(Function, BiFunction, BiConsumer, List)}, but never above {@link #maxRetries}.
+     * {@link #runAsync(Function, BiFunction, BiConsumer, List)}, but never above {@link Config#limit}.
      */
     private int limit;
+    @Nonnull private final Function<Config, Config> configLoader;
+    @Nonnull private Config config;
+    private int configLoaderInvocationCount = 0;
+
     /**
      * The number of successful transactions in a row as called by {@link #runAsync(Function, BiFunction, BiConsumer, List)}.
      */
@@ -186,21 +185,18 @@ public class OnlineIndexer implements AutoCloseable {
     private final boolean syntheticIndex;
 
     @SuppressWarnings("squid:S00107")
-    protected OnlineIndexer(@Nonnull FDBDatabaseRunner runner,
+    OnlineIndexer(@Nonnull FDBDatabaseRunner runner,
                             @Nonnull FDBRecordStore.Builder recordStoreBuilder,
                             @Nonnull Index index, @Nonnull Collection<RecordType> recordTypes,
-                            int maxLimit, int maxRetries, int recordsPerSecond, long progressLogIntervalMillis,
-                            int increaseLimitAfter, boolean syntheticIndex) {
+                            @Nonnull Function<Config, Config> configLoader, @Nonnull Config config,
+                            boolean syntheticIndex) {
         this.runner = runner;
         this.recordStoreBuilder = recordStoreBuilder;
         this.index = index;
         this.recordTypes = recordTypes;
-        this.limit = maxLimit;
-        this.maxLimit = maxLimit;
-        this.maxRetries = maxRetries;
-        this.recordsPerSecond = recordsPerSecond;
-        this.progressLogIntervalMillis = progressLogIntervalMillis;
-        this.increaseLimitAfter = increaseLimitAfter;
+        this.configLoader = configLoader;
+        this.config = config;
+        this.limit = config.maxLimit;
         this.syntheticIndex = syntheticIndex;
         this.recordsRange = computeRecordsRange();
         timeOfLastProgressLogMillis = System.currentTimeMillis();
@@ -221,6 +217,26 @@ public class OnlineIndexer implements AutoCloseable {
             addLogInfo(LogMessageKeys.RANGE_END, end);
         }
     }
+
+    /**
+     * Get the current config parameters of the online indexer.
+     * @return the config parameters of the online indexer
+     */
+    @Nonnull
+    @VisibleForTesting
+    Config getConfig() {
+        return config;
+    }
+
+    /**
+     * Get the number of times the configuration was loaded.
+     * @return the number of times the {@code configLoader} was invoked
+     */
+    @VisibleForTesting
+    int getConfigLoaderInvocationCount() {
+        return configLoaderInvocationCount;
+    }
+
 
     /**
      * Get the current number of records to process in one transaction.
@@ -294,6 +310,23 @@ public class OnlineIndexer implements AutoCloseable {
         return recordStoreBuilder.copyBuilder().setContext(context).openAsync();
     }
 
+    private void loadConfig() {
+        configLoaderInvocationCount++;
+        if (configLoader != null) {
+            config = configLoader.apply(config);
+            if (limit > config.maxLimit) {
+                if (LOGGER.isInfoEnabled()) {
+                    LOGGER.info(
+                            KeyValueLogMessage.build("Decreasing the limit to the new maxLimit.",
+                                    LogMessageKeys.INDEX_NAME, index.getName(),
+                                    LogMessageKeys.LIMIT, limit,
+                                    LogMessageKeys.MAX_LIMIT, config.maxLimit).toString());
+                }
+                limit = config.maxLimit;
+            }
+        }
+    }
+
     @Override
     public void close() {
         runner.close();
@@ -332,41 +365,40 @@ public class OnlineIndexer implements AutoCloseable {
         AtomicInteger tries = new AtomicInteger(0);
         CompletableFuture<R> ret = new CompletableFuture<>();
         AtomicLong toWait = new AtomicLong(FDBDatabaseFactory.instance().getInitialDelayMillis());
-
-        AsyncUtil.whileTrue(() ->
-                runner.runAsync(context -> openRecordStore(context).thenCompose(store -> {
-                    if (!store.isIndexWriteOnly(index)) {
-                        throw new RecordCoreStorageException("Attempted to build readable index",
-                                LogMessageKeys.INDEX_NAME, index.getName(),
-                                recordStoreBuilder.getSubspaceProvider().logKey(), recordStoreBuilder.getSubspaceProvider().toString(context));
-                    }
-                    return function.apply(store);
-                }), handlePostTransaction, onlineIndexerLogMessageKeyValues).handle((value, e) -> {
-                    if (e == null) {
-                        ret.complete(value);
-                        return AsyncUtil.READY_FALSE;
-                    } else {
-                        int currTries = tries.getAndIncrement();
-                        FDBException fdbE = getFDBException(e);
-                        if (currTries < maxRetries && fdbE != null && lessenWorkCodes.contains(fdbE.getCode())) {
-                            if (handleLessenWork != null) {
-                                handleLessenWork.accept(fdbE, onlineIndexerLogMessageKeyValues);
-                            }
-                            long delay = (long)(Math.random() * toWait.get());
-                            toWait.set(Math.min(delay * 2, FDBDatabaseFactory.instance().getMaxDelayMillis()));
-                            return MoreAsyncUtil.delayedFuture(delay, TimeUnit.MILLISECONDS).thenApply(vignore3 -> true);
-                        } else {
-                            return completeExceptionally(ret, e, onlineIndexerLogMessageKeyValues);
+        AsyncUtil.whileTrue(() -> {
+            loadConfig();
+            return runner.runAsync(context -> openRecordStore(context).thenCompose(store -> {
+                if (!store.isIndexWriteOnly(index)) {
+                    throw new RecordCoreStorageException("Attempted to build readable index",
+                            LogMessageKeys.INDEX_NAME, index.getName(),
+                            recordStoreBuilder.getSubspaceProvider().logKey(), recordStoreBuilder.getSubspaceProvider().toString(context));
+                }
+                return function.apply(store);
+            }), handlePostTransaction, onlineIndexerLogMessageKeyValues).handle((value, e) -> {
+                if (e == null) {
+                    ret.complete(value);
+                    return AsyncUtil.READY_FALSE;
+                } else {
+                    int currTries = tries.getAndIncrement();
+                    FDBException fdbE = getFDBException(e);
+                    if (currTries < config.maxRetries && fdbE != null && lessenWorkCodes.contains(fdbE.getCode())) {
+                        if (handleLessenWork != null) {
+                            handleLessenWork.accept(fdbE, onlineIndexerLogMessageKeyValues);
                         }
+                        long delay = (long)(Math.random() * toWait.get());
+                        toWait.set(Math.min(delay * 2, FDBDatabaseFactory.instance().getMaxDelayMillis()));
+                        return MoreAsyncUtil.delayedFuture(delay, TimeUnit.MILLISECONDS).thenApply(vignore3 -> true);
+                    } else {
+                        return completeExceptionally(ret, e, onlineIndexerLogMessageKeyValues);
                     }
-                }).thenCompose(Function.identity()), runner.getExecutor())
-                .whenComplete((vignore, e) -> {
-                    if (e != null) {
-                        // Just update ret and ignore the returned future.
-                        completeExceptionally(ret, e, onlineIndexerLogMessageKeyValues);
-                    }
-                });
-
+                }
+            }).thenCompose(Function.identity());
+        }, runner.getExecutor()).whenComplete((vignore, e) -> {
+            if (e != null) {
+                // Just update ret and ignore the returned future.
+                completeExceptionally(ret, e, onlineIndexerLogMessageKeyValues);
+            }
+        });
         return ret;
     }
 
@@ -420,10 +452,10 @@ public class OnlineIndexer implements AutoCloseable {
     }
 
     private void tryToIncreaseLimit(@Nullable Throwable exception) {
-        if (increaseLimitAfter > 0) {
+        if (config.increaseLimitAfter > 0) {
             if (exception == null) {
                 successCount++;
-                if (successCount >= increaseLimitAfter && limit < maxLimit) {
+                if (successCount >= config.increaseLimitAfter && limit < config.maxLimit) {
                     increaseLimit();
                 }
             } else {
@@ -433,7 +465,7 @@ public class OnlineIndexer implements AutoCloseable {
     }
 
     private void increaseLimit() {
-        limit = Math.min(maxLimit, Math.max(limit + 1, (4 * limit) / 3));
+        limit = Math.min(config.maxLimit, Math.max(limit + 1, (4 * limit) / 3));
         if (LOGGER.isInfoEnabled()) {
             LOGGER.info(KeyValueLogMessage.of("Re-increasing limit of online index build",
                             LogMessageKeys.INDEX_NAME, index.getName(),
@@ -660,7 +692,7 @@ public class OnlineIndexer implements AutoCloseable {
                                                         Tuple startTuple, Tuple endTuple, Tuple realEnd,
                                                         Throwable ex) {
         final RuntimeException unwrappedEx = ex == null ? null : runner.getDatabase().mapAsyncToSyncException(ex);
-        long toWait = (recordsPerSecond == UNLIMITED) ? 0 : 1000 * limit / recordsPerSecond;
+        long toWait = (config.recordsPerSecond == UNLIMITED) ? 0 : 1000 * limit / config.recordsPerSecond;
         if (unwrappedEx == null) {
             if (realEnd != null && !realEnd.equals(endTuple)) {
                 // We didn't make it to the end. Continue on to the next item.
@@ -697,9 +729,9 @@ public class OnlineIndexer implements AutoCloseable {
 
     private void maybeLogBuildProgress(SubspaceProvider subspaceProvider, Tuple startTuple, Tuple endTuple, Tuple realEnd) {
         if (LOGGER.isInfoEnabled()
-                && (progressLogIntervalMillis > 0
-                    && System.currentTimeMillis() - timeOfLastProgressLogMillis > progressLogIntervalMillis)
-                || progressLogIntervalMillis == 0) {
+                && (config.progressLogIntervalMillis > 0
+                    && System.currentTimeMillis() - timeOfLastProgressLogMillis > config.progressLogIntervalMillis)
+                || config.progressLogIntervalMillis == 0) {
             LOGGER.info(KeyValueLogMessage.of("Built Range",
                             LogMessageKeys.INDEX_NAME, index.getName(),
                             LogMessageKeys.INDEX_VERSION, index.getLastModifiedVersion(),
@@ -1158,6 +1190,183 @@ public class OnlineIndexer implements AutoCloseable {
     }
 
     /**
+     * A holder for the mutable configuration parameters needed to rebuild an online index. These parameters are
+     * designed to be safe to be updated while a build is running.
+     */
+    @API(API.Status.UNSTABLE)
+    public static class Config {
+        private final int maxLimit;
+        private final int maxRetries;
+        private final int recordsPerSecond;
+        private final long progressLogIntervalMillis;
+        private final int increaseLimitAfter;
+
+        private Config(int maxLimit, int maxRetries, int recordsPerSecond, long progressLogIntervalMillis, int increaseLimitAfter) {
+            this.maxLimit = maxLimit;
+            this.maxRetries = maxRetries;
+            this.recordsPerSecond = recordsPerSecond;
+            this.progressLogIntervalMillis = progressLogIntervalMillis;
+            this.increaseLimitAfter = increaseLimitAfter;
+        }
+
+        /**
+         * Get the maximum number of records to process in one transaction.
+         * @return the maximum number of records to process in one transaction
+         */
+        public int getMaxLimit() {
+            return maxLimit;
+        }
+
+        /**
+         * Get the maximum number of times to retry a single range rebuild.
+         * @return the maximum number of times to retry a single range rebuild
+         */
+        public int getMaxRetries() {
+            return maxRetries;
+        }
+
+        /**
+         * Get the maximum number of records to process in a single second.
+         * @return the maximum number of records to process in a single second
+         */
+        public int getRecordsPerSecond() {
+            return recordsPerSecond;
+        }
+
+        /**
+         * Get the minimum time between successful progress logs when building across transactions.
+         * Negative will not log at all, 0 will log after every commit.
+         * @return the minimum time between successful progress logs in milliseconds
+         */
+        public long getProgressLogIntervalMillis() {
+            return progressLogIntervalMillis;
+        }
+
+        /**
+         * Get the number of successful range builds before re-increasing the number of records to process in a single
+         * transaction.
+         * By default this is {@link #DO_NOT_RE_INCREASE_LIMIT}, which means it will not re-increase after successes.
+         * @return the number of successful range builds before increasing the number of records processed in a single
+         * transaction
+         */
+        public int getIncreaseLimitAfter() {
+            return increaseLimitAfter;
+        }
+
+        @Nonnull
+        public static Builder newBuilder() {
+            return new Builder();
+        }
+
+        /**
+         * To create a builder for the given config.
+         * @return a {@link Config.Builder}
+         */
+        @Nonnull
+        public Builder toBuilder() {
+            return Config.newBuilder()
+                    .setMaxLimit(this.maxLimit)
+                    .setIncreaseLimitAfter(this.increaseLimitAfter)
+                    .setProgressLogIntervalMillis(this.progressLogIntervalMillis)
+                    .setRecordsPerSecond(this.recordsPerSecond)
+                    .setMaxRetries(this.maxRetries);
+        }
+
+        /**
+         * A builder for {@link Config}. These are the mutable configuration parameters used while building indexes and are
+         * designed to be safe to be updated while a build is running.
+         */
+        @API(API.Status.UNSTABLE)
+        public static class Builder {
+            private int maxLimit = DEFAULT_LIMIT;
+            private int maxRetries = DEFAULT_MAX_RETRIES;
+            private int recordsPerSecond = DEFAULT_RECORDS_PER_SECOND;
+            private long progressLogIntervalMillis = DEFAULT_PROGRESS_LOG_INTERVAL;
+            private int increaseLimitAfter = DO_NOT_RE_INCREASE_LIMIT;
+
+            protected Builder() {
+
+            }
+
+            /**
+             * Set the maximum number of records to process in one transaction.
+             *
+             * The default limit is {@link #DEFAULT_LIMIT} = {@value #DEFAULT_LIMIT}.
+             * @param limit the maximum number of records to process in one transaction
+             * @return this builder
+             */
+            @Nonnull
+            public Builder setMaxLimit(int limit) {
+                this.maxLimit = limit;
+                return this;
+            }
+
+            /**
+             * Set the maximum number of times to retry a single range rebuild.
+             *
+             * The default number of retries is {@link #DEFAULT_MAX_RETRIES} = {@value #DEFAULT_MAX_RETRIES}.
+             * @param maxRetries the maximum number of times to retry a single range rebuild
+             * @return this builder
+             */
+            @Nonnull
+            public Builder setMaxRetries(int maxRetries) {
+                this.maxRetries = maxRetries;
+                return this;
+            }
+
+            /**
+             * Set the maximum number of records to process in a single second.
+             *
+             * The default number of retries is {@link #DEFAULT_RECORDS_PER_SECOND} = {@value #DEFAULT_RECORDS_PER_SECOND}.
+             * @param recordsPerSecond the maximum number of records to process in a single second
+             * @return this builder
+             */
+            @Nonnull
+            public Builder setRecordsPerSecond(int recordsPerSecond) {
+                this.recordsPerSecond = recordsPerSecond;
+                return this;
+            }
+
+
+            /**
+             * Set the minimum time between successful progress logs when building across transactions.
+             * Negative will not log at all, 0 will log after every commit.
+             *
+             * @param progressLogIntervalMillis the number of milliseconds to wait between successful logs
+             * @return this builder
+             */
+            @Nonnull
+            public Builder setProgressLogIntervalMillis(long progressLogIntervalMillis) {
+                this.progressLogIntervalMillis = progressLogIntervalMillis;
+                return this;
+            }
+
+            /**
+             * Set the number of successful range builds before re-increasing the number of records to process in a single
+             * transaction. The number of records to process in a single transaction will never go above {@link #getMaxLimit()}.
+             * By default this is {@link #DO_NOT_RE_INCREASE_LIMIT}, which means it will not re-increase after successes.
+             * @param increaseLimitAfter the number of successful range builds before increasing the number of records
+             * processed in a single transaction
+             * @return this builder
+             */
+            @Nonnull
+            public Builder setIncreaseLimitAfter(int increaseLimitAfter) {
+                this.increaseLimitAfter = increaseLimitAfter;
+                return this;
+            }
+
+            /**
+             * Build a {@link Config}.
+             * @return a new Config object needed by {@link OnlineIndexer}
+             */
+            @Nonnull
+            public Config build() {
+                return new Config(maxLimit, maxRetries, recordsPerSecond, progressLogIntervalMillis, increaseLimitAfter);
+            }
+        }
+    }
+
+    /**
      * Builder for {@link OnlineIndexer}.
      *
      * <pre><code>
@@ -1180,6 +1389,8 @@ public class OnlineIndexer implements AutoCloseable {
         @Nullable
         protected Collection<RecordType> recordTypes;
 
+        @Nonnull
+        protected Function<Config, Config> configLoader = old -> old;
         protected int limit = DEFAULT_LIMIT;
         protected int maxRetries = DEFAULT_MAX_RETRIES;
         protected int recordsPerSecond = DEFAULT_RECORDS_PER_SECOND;
@@ -1323,6 +1534,31 @@ public class OnlineIndexer implements AutoCloseable {
         }
 
         /**
+         * Get the function used by the online indexer to load the config parameters on fly.
+         * @return the function
+         */
+        @Nullable
+        public Function<Config, Config> getConfigLoader() {
+            return configLoader;
+        }
+
+        /**
+         * Set the function used by the online indexer to load the mutable configuration parameters on fly.
+         *
+         * <p>
+         * The loader is given the current configuration as input at the beginning of each transaction and
+         * should produce the configuration to use in the next transaction.
+         * </p>
+         * @param configLoader the function
+         * @return this builder
+         */
+        @Nonnull
+        public Builder setConfigLoader(@Nonnull Function<Config, Config> configLoader) {
+            this.configLoader = configLoader;
+            return this;
+        }
+
+        /**
          * Get the maximum number of records to process in one transaction.
          * @return the maximum number of records to process in one transaction
          */
@@ -1334,6 +1570,8 @@ public class OnlineIndexer implements AutoCloseable {
          * Set the maximum number of records to process in one transaction.
          *
          * The default limit is {@link #DEFAULT_LIMIT} = {@value #DEFAULT_LIMIT}.
+         * Note {@link #setConfigLoader(Function)} is the recommended way of loading online index builder's parameters
+         * and the values set by this method will be overwritten if the supplier is set.
          * @param limit the maximum number of records to process in one transaction
          * @return this builder
          */
@@ -1359,6 +1597,8 @@ public class OnlineIndexer implements AutoCloseable {
          * codes, such as {@code transaction_too_large}.
          *
          * The default number of retries is {@link #DEFAULT_MAX_RETRIES} = {@value #DEFAULT_MAX_RETRIES}.
+         * Note {@link #setConfigLoader(Function)} is the recommended way of loading online index builder's parameters
+         * and the values set by this method will be overwritten if the supplier is set.
          * @param maxRetries the maximum number of times to retry a single range rebuild
          * @return this builder
          */
@@ -1380,6 +1620,8 @@ public class OnlineIndexer implements AutoCloseable {
          * Set the maximum number of records to process in a single second.
          *
          * The default number of retries is {@link #DEFAULT_RECORDS_PER_SECOND} = {@value #DEFAULT_RECORDS_PER_SECOND}.
+         * Note {@link #setConfigLoader(Function)} is the recommended way of loading online index builder's parameters
+         * and the values set by this method will be overwritten if the supplier is set.
          * @param recordsPerSecond the maximum number of records to process in a single second.
          * @return this builder
          */
@@ -1550,6 +1792,10 @@ public class OnlineIndexer implements AutoCloseable {
          * Set the number of successful range builds before re-increasing the number of records to process in a single
          * transaction. The number of records to process in a single transaction will never go above {@link #limit}.
          * By default this is {@link #DO_NOT_RE_INCREASE_LIMIT}, which means it will not re-increase after successes.
+         * <p>
+         * Note {@link #setConfigLoader(Function)} is the recommended way of loading online index builder's parameters
+         * and the values set by this method will be overwritten if the supplier is set.
+         * </p>
          * @param increaseLimitAfter the number of successful range builds before increasing the number of records
          * processed in a single transaction
          * @return this builder
@@ -1654,6 +1900,11 @@ public class OnlineIndexer implements AutoCloseable {
          *         regardless of whether the associated transaction was successful or not.
          *     </p></li>
          * </ul>
+         *
+         * <p>
+         * Note {@link #setConfigLoader(Function)} is the recommended way of loading online index builder's parameters
+         * and the values set by this method will be overwritten if the supplier is set.
+         * </p>
          *
          * @param millis the number of milliseconds to wait between successful logs
          * @return this builder
@@ -1764,7 +2015,8 @@ public class OnlineIndexer implements AutoCloseable {
          */
         public OnlineIndexer build() {
             validate();
-            return new OnlineIndexer(runner, recordStoreBuilder, index, recordTypes, limit, maxRetries, recordsPerSecond, progressLogIntervalMillis, increaseLimitAfter, syntheticIndex);
+            Config conf = new Config(limit, maxRetries, recordsPerSecond, progressLogIntervalMillis, increaseLimitAfter);
+            return new OnlineIndexer(runner, recordStoreBuilder, index, recordTypes, configLoader, conf, syntheticIndex);
         }
 
         protected void validate() {